pyproject.toml
[project]
name = "tab-to-pbi"
version = "0.1.0"
description = "Tableau to Power BI migration tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.89.0",
    "jsonschema>=4.26.0",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "pywinauto>=0.6.9",
]

[tool.setuptools.packages.find]
include = ["tab_to_pbi*"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
